requirements.txt
# Application
PySide6-Essentials~=6.8
mosamatic-cli>=1.5
pydicom